+++ src/env/object.rs
@@ -1,5 +1,6 @@
+use rug::Integer;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum Object {
-    Integer(i64)
+    Integer(Integer)
 }
